Bump major version
