+++ src/Custom.Build.props
@@ -1,7 +1,7 @@
 <Project>
 
   <PropertyGroup>
-    <MinVerMinimumMajorMinor>8.1</MinVerMinimumMajorMinor>
+    <MinVerMinimumMajorMinor>9.0</MinVerMinimumMajorMinor>
     <MinVerAutoIncrement>minor</MinVerAutoIncrement>
   </PropertyGroup>
